+++ Src/Roflcopter.Plugin.Tests/test/data/MismatchedFileNames/ClassAndInterfacePair.cs
@@ -1,0 +1,10 @@
+﻿namespace Roflcopter.Sample.MismatchedFileNames
+{
+    public class ClassAndInterfacePair
+    {
+    }
+
+    public interface IClassAndInterfacePair
+    {
+    }
+}