+++ Src/Roflcopter.Plugin.Tests/test/data/MismatchedFileNames/ClassAndInterfacePairWithWrongName.cs
@@ -1,0 +1,10 @@
+﻿namespace Roflcopter.Sample.MismatchedFileNames
+{
+    public class ClassAndInterfacePairWithWrongNameX
+    {
+    }
+
+    public interface IClassAndInterfacePairWithWrongNameX
+    {
+    }
+}