+++ Src/Roflcopter.Plugin.Tests/test/data/MismatchedFileNames/InterfaceAndClassPair.cs
@@ -1,0 +1,10 @@
+﻿namespace Roflcopter.Sample.MismatchedFileNames
+{
+    public interface IInterfaceAndClassPair
+    {
+    }
+
+    public class InterfaceAndClassPair
+    {
+    }
+}