+++ Src/Roflcopter.Plugin.Tests/test/data/MismatchedFileNames/MultipleClasses.cs
@@ -1,0 +1,14 @@
+﻿namespace Roflcopter.Sample.MismatchedFileNames
+{
+    public class MultipleClassesA
+    {
+    }
+
+    public class MultipleClassesB
+    {
+    }
+
+    public class MultipleClassesC
+    {
+    }
+}