+++ Src/Roflcopter.Plugin/MismatchedFileNames/MismatchedFileNameHighlightingElementProblemAnalyzer.cs
@@ -1,4 +1,5 @@
 ﻿using System.IO;
+using JetBrains.Annotations;
 using JetBrains.ReSharper.Daemon.Stages.Dispatcher;
 using JetBrains.ReSharper.Feature.Services.Daemon;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
@@ -10,14 +11,33 @@ namespace Roflcopter.Plugin.MismatchedFileNames
     [ElementProblemAnalyzer(
         typeof(ICSharpFile),
         HighlightingTypes = new[] { typeof(MismatchedFileNameHighlighting) })]
-    public class MismatchedFileNameHighlightingElementProblemAnalyzer : ElementProblemAnalyzer<ICSharpFile>
+    public class MismatchedFileNameElementProblemAnalyzer : ElementProblemAnalyzer<ICSharpFile>
     {
         protected override void Run(
             ICSharpFile file,
             ElementProblemAnalyzerData data,
             IHighlightingConsumer consumer)
         {
-            var typeDeclarations = new LocalList<ITypeDeclaration>();
+            var tolLevelTypeDeclarations = GetTolLevelTypeDeclarations(file);
+
+            var mainTypeDeclaration = FindMainTypeDeclaration(tolLevelTypeDeclarations);
+
+            if (mainTypeDeclaration != null)
+            {
+                var psiSourceFile = file.GetSourceFile().NotNull("file.GetSourceFile() != null");
+
+                var fileNameWithoutExtension = Path.GetFileNameWithoutExtension(psiSourceFile.Name);
+
+                if (fileNameWithoutExtension != mainTypeDeclaration.DeclaredName)
+                {
+                    consumer.AddHighlighting(new MismatchedFileNameHighlighting(mainTypeDeclaration, psiSourceFile.Name));
+                }
+            }
+        }
+
+        private static LocalList<ITypeDeclaration> GetTolLevelTypeDeclarations(IFile file)
+        {
+            var result = new LocalList<ITypeDeclaration>();
 
             var descendantsEnumerator = file.Descendants();
 
@@ -25,24 +45,33 @@ namespace Roflcopter.Plugin.MismatchedFileNames
             {
                 if (descendantsEnumerator.Current is ITypeDeclaration typeDeclaration)
                 {
-                    typeDeclarations.Add(typeDeclaration);
+                    result.Add(typeDeclaration);
                     descendantsEnumerator.SkipThisNode();
                 }
             }
-
-            if (typeDeclarations.Count == 1)
-            {
-                var typeDeclaration = typeDeclarations[0];
-
-                var psiSourceFile = file.GetSourceFile().NotNull("file.GetSourceFile() != null");
-
-                var fileNameWithoutExtension = Path.GetFileNameWithoutExtension(psiSourceFile.Name);
-
-                if (fileNameWithoutExtension != typeDeclaration.DeclaredName)
-                {
-                    consumer.AddHighlighting(new MismatchedFileNameHighlighting(typeDeclaration, psiSourceFile.Name));
-                }
-            }
+            return result;
         }
+
+        [CanBeNull]
+        private static ITypeDeclaration FindMainTypeDeclaration(LocalList<ITypeDeclaration> typeDeclarations)
+        {
+            if (typeDeclarations.Count == 1)
+                return typeDeclarations[0];
+
+            if (typeDeclarations.Count == 2)
+            {
+                if (IsDeclarationPair(typeDeclarations[0], typeDeclarations[1]))
+                    return typeDeclarations[0];
+
+                if (IsDeclarationPair(typeDeclarations[1], typeDeclarations[0]))
+                    return typeDeclarations[1];
+            }
+
+            return null;
+        }
+
+        [Pure]
+        private static bool IsDeclarationPair(ITypeDeclaration mainTypeDeclaration, ITypeDeclaration otherTypeDeclaration) =>
+            "I" + mainTypeDeclaration.DeclaredName == otherTypeDeclaration.DeclaredName;
     }
 }